:art: Refactor all styling code out of AddDeckForm
